FIX ALL THE THINGS \o/ in Equipment module

+++ game/src/snake/equipment/IEquipmentCreator.java
@@ -2,5 +2,5 @@ package snake.equipment;
 
 public interface IEquipmentCreator 
 {
-	public IEquipment create(int x, int y);
+	public IEquipment create(float x, float y);
 }
